+++ tests/AppserverIo/Http/Functional/DigestAuthenticationTest.php
@@ -21,6 +21,7 @@
 namespace AppserverIo\Http\Functional;
 
 use AppserverIo\Http\Authentication\DigestAuthentication;
+use AppserverIo\Psr\HttpMessage\Protocol;
 
 /**
  * Class for testing the digest authentication feature
@@ -87,11 +88,26 @@ class DigestAuthenticationTest extends \PHPUnit_Framework_TestCase
      * Tests if we are deflected using an invalid auth string and the default adapter
      *
      * @return void
+     * @expectedException AppserverIo\Http\Authentication\AuthenticationException
      */
     public function testInvalidAuthDefaultAdapter()
     {
-        $this->testClass->init($this->getInvalidAuthString(), 'GET');
-        $this->assertFalse($this->testClass->authenticate());
+
+        // prepare the mock request instance
+        $mockRequest = $this->getMock('AppserverIo\Psr\HttpMessage\RequestInterface');
+        $mockRequest->expects($this->once())
+                    ->method('getMethod')
+                    ->willReturn(Protocol::METHOD_GET);
+        $mockRequest->expects($this->once())
+                    ->method('getHeader')
+                    ->with(Protocol::HEADER_AUTHORIZATION)
+                    ->willReturn($this->getInvalidAuthString());
+
+        // prepare the mock response instance
+        $mockResponse = $this->getMock('AppserverIo\Psr\HttpMessage\ResponseInterface');
+
+        $this->testClass->init($mockRequest, $mockResponse);
+        $this->testClass->authenticate($mockResponse);
     }
 
     /**
@@ -101,7 +117,21 @@ class DigestAuthenticationTest extends \PHPUnit_Framework_TestCase
      */
     public function testValidAuthDefaultAdapter()
     {
-        $this->testClass->init($this->getValidAuthString(), 'GET');
-        $this->assertTrue($this->testClass->authenticate());
+
+        // prepare the mock request instance
+        $mockRequest = $this->getMock('AppserverIo\Psr\HttpMessage\RequestInterface');
+        $mockRequest->expects($this->once())
+                    ->method('getMethod')
+                    ->willReturn(Protocol::METHOD_GET);
+        $mockRequest->expects($this->once())
+                    ->method('getHeader')
+                    ->with(Protocol::HEADER_AUTHORIZATION)
+                    ->willReturn($this->getValidAuthString());
+
+        // prepare the mock response instance
+        $mockResponse = $this->getMock('AppserverIo\Psr\HttpMessage\ResponseInterface');
+
+        $this->testClass->init($mockRequest, $mockResponse);
+        $this->assertNull($this->testClass->authenticate($mockResponse));
     }
 }